Add bike to access page
Add a new section to promote bike as a way to go to Skema.
Move 'by car' down in the list, we'd rather promote bike and public transportation first.

I don't have the same icon size for bike so I'm cheating with CSS background size. It is not perfect
but likely good enough :).